media: dvb-core: Fix use-after-free due to race condition at dvb_ca_en50221
jira ROC-3033
cve CVE-2022-45919
commit 280a8ab81733da8bc442253c700a52c4c0886ffd

If the device node of dvb_ca_en50221 is open() and the
device is disconnected, a UAF may occur when calling
close() on the device node.

The root cause is that wake_up() and wait_event() for
dvbdev->wait_queue are not implemented.

So implement wait_event() function in dvb_ca_en50221_release()
and add 'remove_mutex' which prevents race condition
for 'ca->exit'.

[mchehab: fix a checkpatch warning]

    Link: https://lore.kernel.org/linux-media/20221121063308.GA33821@ubuntu
    Signed-off-by: Hyunwoo Kim <v4bel@theori.io>
    Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
    (cherry picked from commit 280a8ab81733da8bc442253c700a52c4c0886ffd)
Signed-off-by: Ronnie Sahlberg <rsahlberg@ciq.com>