md: partially revert "md/raid6: use valid sector values to determine if an I/O should wait on the reshape"
JIRA: https://issues.redhat.com/browse/RHEL-24489
Upstream Status: RHEL Only

Revert the max sector calculation from commit c467e97f079f ("md/raid6:
use valid sector values to determine if an I/O should wait on the
reshape"). It can cause a deadlock during dm-raid reshaping.

This fix was originally posted to dm-devel by Mikulas Patocka, but not
accepted upstream.
https://lore.kernel.org/dm-devel/b725da99-d649-6f1d-af82-c3e482f7f6e@redhat.com/

The upstream fix for this issue is still under development and involves
significant changes to the MD code. There isn't sufficient time in the
rhel-9.4 schedule to wait for the upstream fix to be finalized and the
backported code to be tested thoroughly. This RHEL only change will be
removed when then final solution is backported, likely in rhel-9.5.

Co-authored-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>