Revert "arm64: efi: Use SMBIOS processor version to key off Ampere quirk"
JIRA: https://issues.redhat.com/browse/RHEL-23382

Upstream-status: RHEL only

This reverts commit a496c48b33aaa02c17c7122b185a79befc71d60e.  It looks like
there are other systems that fail if SetVirtualAddressMap() is not called,
so just revert the patch that made it optional.

This patch reverts some of the checks that were introduced to work around
firmware bugs; ultimately, the no-SetVirtualAddressMap() path will
become the only one unless "novamap" is passed, therefore it is not
necessary to have these quirks.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

[RHEL-only: revert of buggy commit]