Revert "arm64: efi: Force the use of SetVirtualAddressMap() on Altra machines"
JIRA: https://issues.redhat.com/browse/RHEL-23382

Upstream-status: RHEL only

This reverts commit d6a96b09b078b0382e14a206142c8f063901cd4f.  It looks like
there are other systems that fail if SetVirtualAddressMap() is not called,
so just revert the patch that made it optional.

This patch reverts the remaining checks that were introduced to work
around firmware bugs; ultimately, the no-SetVirtualAddressMap() path
will become the only one unless "novamap" is passed, therefore it is
not necessary to have these quirks.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

[RHEL-only: revert of buggy commit]