regmap: mmio: Allow passing an empty config->reg_stride
JIRA: https://issues.redhat.com/browse/RHEL-26183

commit e12ff28764937dd58c8613f16065da60da149048
Author: Maxime Chevallier <maxime.chevallier@bootlin.com>
Date: Fri, 12 May 2023 10:28:19 +0000

Regmap's stride is used for MMIO regmaps to check the correctness of
reg_width. However, it's acceptable to pass an empty config->reg_stride,
in that case the actual stride used is 1.

There are valid cases now to pass an empty stride, when using
down/upshifting of register address. In this case, the stride value
loses its sense, so ignore the reg_width when the stride isn't set.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com
Link: https://lore.kernel.org/r/20230511142735.316445-1-maxime.chevallier@bootlin.com
Signed-off-by: Mark Brown <broonie@kernel.org
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>