config: Disabling CONFIG_INTEL_MEI_GSC_PROXY
Jira: INTERNAL
Upstream Status: RHEL only

Disabling CONFIG_INTEL_MEI_GSC_PROXY because this config requires to
enable other configs like CONFIG_INTEL_MEI and CONFIG_INTEL_MEI_ME. If
it is required to automotive, it can be enabled again in the future
using an MR for this. For further reference in CS9, see MR!3585.

Signed-off-by: Julio Faracco <jfaracco@redhat.com>