gitlab-ci: Remove the 'check' phase, add a 'deploy' phase. Automatically rebuild info/*.
JIRA: INTERNAL

motivation: Checking if a commiter has run 'make' by running 'make'
(twice) is redundant. Instead, deploy the artifacts of make the first
time around.

This approach simplifies documentation contributions. 'make' will
already fail and exit if a genuine linting error occurs.

Furthermore, run 'make' from the root of the directory, rather than
using the makefile located in 'scripts/'.

Additionally, make sure the `deploy` phase of the pipeline only runs in the
main branch of the parent repo. Pages is already deployed in the
ci-check, so no need to run this in every fork+branch.

Finally, cleanup the following:
- unused scripts/ artifiacts in 'build:'
- unused 'pages:' '-when:never' label.

Signed-off-by: Derek Barbosa <debarbos@redhat.com>