redhat/configs: aarch64: enable SCMI pmdomain driver
JIRA: https://issues.redhat.com/browse/RHEL-26434
Upstream-status: RHEL-only

- CONFIG_ARM_SCMI_PERF_DOMAIN -> y

This driver is newly added to RHEL, and the corresponding Kconfig option
is new as well. Set a default value for the new option across aarch64 to
avoid an error while building the kernel configs. The default value is
aligned with kernel-ark and enables the driver, even though there is no
immediate plan to use the driver in RHEL. The target platforms for this
driver are some of the automotive boards supported in main-automotive.

Signed-off-by: Radu Rendec <rrendec@redhat.com>