ACPI: acpi_pad: Rename ACPI device from device to adev
JIRA: https://issues.redhat.com/browse/RHEL-26871
Conflicts:
	drivers/acpi/acpi_pad.c - minor context
differences. had to replace device with adev in
a few functions.

commit 5ccd40c5c7ada69d7465478336eb50ea93d54d6d
Author: Michal Wilczynski <michal.wilczynski@intel.com>
Date: Tue, 24 Oct 2023 20:54:29 +0000

Since transformation from ACPI driver to platform driver there are two
devices on which the driver operates - ACPI device and platform device.
For the sake of reader this calls for the distinction in their naming,
to avoid confusion. Rename device to adev, as corresponding
platform device is called pdev.

Signed-off-by: Michal Wilczynski <michal.wilczynski@intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>