ACPI: processor: Fix uninitialized access of buf in acpi_set_pdc_bits()
JIRA: https://issues.redhat.com/browse/RHEL-26871

commit b2eb3e67ee68dee9c0555466dfa8d7f0ffcc00db
Author: Michal Wilczynski <michal.wilczynski@intel.com>
Date: Mon, 18 Sep 2023 12:16:16 +0000

A bug was introduced during unification of setting CAP_SMP_T_SWCOORD for
the _PDC and _OSC methods.

The third u32 in the buffer is never cleared before setting bits on it.
The memory is not guaranteed to be zero as it was allocated by kmalloc()
instead of kzalloc().

Fix this by initializing the third u32 in the buffer to 0.

Fixes: b9e8d0168a7a ("ACPI: processor: Set CAP_SMP_T_SWCOORD in arch_acpi_set_proc_cap_bits()")
Signed-off-by: Michal Wilczynski <michal.wilczynski@intel.com>
[ rjw: Subject and changelog edits ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>