ovl: fail on invalid uid/gid mapping at copy up
jira: ROC-3260
cve: CVE-2023-0386
commit: 4f11ada10d0ad3fd53e2bd67806351de63a4f9c3

If st_uid/st_gid doesn't have a mapping in the mounter's user_ns, then
copy-up should fail, just like it would fail if the mounter task was doing
the copy using "cp -a".

There's a corner case where the "cp -a" would succeed but copy up fail: if
there's a mapping of the invalid uid/gid (65534 by default) in the user
namespace.  This is because stat(2) will return this value if the mapping
doesn't exist in the current user_ns and "cp -a" will in turn be able to
create a file with this uid/gid.

This behavior would be inconsistent with POSIX ACL's, which return -1 for
invalid uid/gid which result in a failed copy.

For consistency and simplicity fail the copy of the st_uid/st_gid are
invalid.

    Fixes: 459c7c565ac3 ("ovl: unprivieged mounts")
    Cc: <stable@vger.kernel.org> # v5.11
    Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
    Reviewed-by: Christian Brauner <brauner@kernel.org>
    Reviewed-by: Seth Forshee <sforshee@kernel.org>
    (cherry picked from commit 4f11ada10d0ad3fd53e2bd67806351de63a4f9c3)

Signed-off-by: David Gomez <dgomez@ciq.com>
Signed-off-by: Jonathan Maple <jmaple@ciq.com>