thermal/drivers/ti-soc: Switch to new of API
JIRA: https://issues.redhat.com/browse/RHEL-26871

commit 2cf3c72a3ffba080b8188a07c19514cd43df6097
Author: Daniel Lezcano <daniel.lezcano@linexp.org>
Date:   Fri Aug 5 00:43:40 2022 +0200

The thermal OF code has a new API allowing to migrate the OF
initialization to a simpler approach. The ops are no longer device
tree specific and are the generic ones provided by the core code.

Convert the ops to the thermal_zone_device_ops format and use the new
API to register the thermal zone with these generic ops.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linexp.org>
Link: https://lore.kernel.org/r/20220804224349.1926752-25-daniel.lezcano@linexp.org
Acked-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>