thermal: core: Rework .get_trend() thermal zone callback
JIRA: https://issues.redhat.com/browse/RHEL-26871

commit 8289d810ea85755a9d22f75785806cb34eecd5e5
Author: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Date:   Tue Aug 22 13:40:06 2023 +0200

Passing a struct thermal_trip pointer instead of a trip index to the
.get_trend() thermal zone callback allows one of its 2 implementations,
the thermal_get_trend() function in the ACPI thermal driver, to be
simplified quite a bit, and the other implementation of it in the
ti-soc-thermal driver does not even use the relevant callback argument.

For this reason, change the .get_trend() thermal zone callback
definition and adjust the related code accordingly.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>