ACPI: thermal: Do not use trip indices for cooling device binding
JIRA: https://issues.redhat.com/browse/RHEL-26871

commit d5ea889246b112e228433a5f27f57af90ca0c1fb
Author: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Date: Thu, 28 Sep 2023 19:27:19 +0000

Rearrange the ACPI thermal driver's callback functions used for cooling
device binding and unbinding, acpi_thermal_bind_cooling_device() and
acpi_thermal_unbind_cooling_device(), respectively, so that they use trip
pointers instead of trip indices which is more straightforward and allows
the driver to become independent of the ordering of trips in the thermal
zone structure.

The general functionality is not expected to be changed.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>