igb: simplify pci ops declaration
Author: Jesse Brandeburg <jesse.brandeburg@intel.com>

The igb driver was pre-declaring tons of functions just so that it could
have an early declaration of the pci_driver struct.

Delete a bunch of the declarations and move the struct to the bottom of the
file, after all the functions are declared.

Reviewed-by: Alan Brady <alan.brady@intel.com>
Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Reviewed-by: Maciej Fijalkowski <maciej.fijalkowski@intel.com>
Tested-by: Pucha Himasekhar Reddy <himasekharx.reddy.pucha@intel.com> (A Contingent worker at Intel)
Signed-off-by: Tony Nguyen <anthony.l.nguyen@intel.com>
(cherry picked from commit 47220a1e0b701d144ef20131f31566cf0815095f)
JIRA: https://issues.redhat.com/browse/RHEL-17487
Signed-off-by: Corinna Vinschen <vinschen@redhat.com>