arm: perf: Remove inlines from arm_pmuv3.c
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit 9343c790e6de7edd2bab17572832f4f21c748dc2
Author: James Clark <james.clark@arm.com>
Date: Mon Dec 11 16:13:13 2023 +0000

description
===========
These are all static and in one compilation unit so the inline has no
effect on the binary. Except if FTRACE is enabled, then 3 functions
which were already not inlined now get the nops added which allows them
to be traced.

    Signed-off-by: James Clark <james.clark@arm.com>
    Link: https://lore.kernel.org/r/20231211161331.1277825-2-james.clark@arm.com
    Signed-off-by: Will Deacon <will@kernel.org>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>