perf pmus: Simplify perf_pmus__find_core_pmu()
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit 105e5b433e5c743bcdb4956b00b772dfddbee455
Author: James Clark <james.clark@arm.com>
Date: Wed Sep 13 16:33:49 2023 +0100

description
===========
Currently the while loop always either exits on the first iteration with
a core PMU, or exits with NULL on heterogeneous systems or when not all
CPUs are online.

Both of the latter behaviors are undesirable for platforms other than
Arm so simplify it to always return the first core PMU, or NULL if none
exist.

This behavior was depended on by the Arm version of
pmu_metrics_table__find(), so the logic has been moved there instead.

    Signed-off-by: James Clark <james.clark@arm.com>
    Suggested-by: Ian Rogers <irogers@google.com>
    Reviewed-by: Ian Rogers <irogers@google.com>
    Reviewed-by: John Garry <john.g.garry@oracle.com>
    Cc: Ravi Bangoria <ravi.bangoria@amd.com>
    Cc: Eduard Zingerman <eddyz87@gmail.com>
    Cc: Will Deacon <will@kernel.org>
    Cc: Leo Yan <leo.yan@linaro.org>
    Cc: Mike Leach <mike.leach@linaro.org>
    Cc: Jing Zhang <renyu.zj@linux.alibaba.com>
    Cc: Haixin Yu <yuhaixin.yhx@linux.alibaba.com>
    Cc: Kan Liang <kan.liang@linux.intel.com>
    Cc: linux-arm-kernel@lists.infradead.org
    Link: https://lore.kernel.org/r/20230913153355.138331-3-james.clark@arm.com
    Signed-off-by: Namhyung Kim <namhyung@kernel.org>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>