perf symbols: Add 'intel_idle_ibrs' to the list of idle symbols
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit 29a2fd7c72b3bd8b67bcee363f8068275a4ef351
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Oct 5 10:29:38 2023 -0300

description
===========
This is a longstanding to do list entry: we need a way to see that a
sample took place while in idle state, as the current way to do it is
to infer that by the name of the functions that in such state have
more samples, IOW: a hack.

Maybe we can do flip a bit in samples that take place inside the
enter/exit idle section in do_idle()?

But till then, add one more :-\

    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>
    Acked-by: Namhyung Kim <namhyung@kernel.org>
    Cc: Frederic Weisbecker <frederic@kernel.org>
    Link: https://lore.kernel.org/r/ZR66Qgbcltt+zG7F@kernel.org
    Signed-off-by: Namhyung Kim <namhyung@kernel.org>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>