perf cs-etm: Fix incorrect or missing decoder for raw trace
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit a16afcc58a8c5ebc65c852faf001f8f61f05e4ef
Author: Besar Wicaksono <bwicaksono@nvidia.com>
Date: Tue Oct 10 18:48:03 2023 -0500

description
===========
The decoder creation for raw trace uses metadata from the first CPU.
On per-cpu mode, this metadata is incorrectly used for every decoder.
On per-process/per-thread traces, the first CPU is CPU0. If CPU0 trace
is not enabled, its metadata will be marked unused and the decoder is
not created. Perf report dump skips the decoding part because the
decoder is missing.

To fix this, use metadata of the CPU associated with sample object.

    Signed-off-by: Besar Wicaksono <bwicaksono@nvidia.com>
    Reviewed-by: James Clark <james.clark@arm.com>
    Cc: suzuki.poulose@arm.com
    Cc: mike.leach@linaro.org
    Cc: jonathanh@nvidia.com
    Cc: rwiley@nvidia.com
    Cc: treding@nvidia.com
    Cc: vsethi@nvidia.com
    Cc: ywan@nvidia.com
    Cc: linux-arm-kernel@lists.infradead.org
    Cc: coresight@lists.linaro.org
    Cc: linux-tegra@vger.kernel.org
    Link: https://lore.kernel.org/r/20231010234803.5419-1-bwicaksono@nvidia.com
    Signed-off-by: Namhyung Kim <namhyung@kernel.org>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>