perf cs-etm: Validate timestamp tracing in per-thread mode
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit f8ccc2d5cc6516b019bcf8e361ae2a380cb36019
Author: Leo Yan <leo.yan@linaro.org>
Date: Sat Oct 14 15:41:58 2023 +0800

description
===========
So far, it's impossible to validate timestamp trace in Arm CoreSight when
the perf is in the per-thread mode.  E.g. for the command:

  perf record -e cs_etm/timestamp/ --per-thread -- ls

The command enables config 'timestamp' for 'cs_etm' event in the
per-thread mode.  In this case, the function cs_etm_validate_config()
directly bails out and skips validation.

Given profiled process can be scheduled on any CPUs in the per-thread
mode, this patch validates timestamp tracing for all CPUs when detect
the CPU map is empty.

    Signed-off-by: Leo Yan <leo.yan@linaro.org>
    Reviewed-by: James Clark <james.clark@arm.com>
    Acked-by: Suzuki K Poulose <suzuki.poulose@arm.com>
    Cc: Will Deacon <will@kernel.org>
    Cc: Mike Leach <mike.leach@linaro.org>
    Cc: John Garry <john.g.garry@oracle.com>
    Cc: linux-arm-kernel@lists.infradead.org
    Cc: coresight@lists.linaro.org
    Link: https://lore.kernel.org/r/20231014074159.1667880-2-leo.yan@linaro.org
    Signed-off-by: Namhyung Kim <namhyung@kernel.org>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>