perf lock contention: Check race in tstamp elem creation
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit 6a070573f290f99a6129ac3e13b9df521a1a65de
Author: Namhyung Kim <namhyung@kernel.org>
Date: Fri Oct 20 13:47:40 2023 -0700

description
===========
When pelem is NULL, it'd create a new entry with zero data.  But it
might be preempted by IRQ/NMI just before calling bpf_map_update_elem()
then there's a chance to call it twice for the same pid.  So it'd be
better to use BPF_NOEXIST flag and check the return value to prevent
the race.

    Signed-off-by: Namhyung Kim <namhyung@kernel.org>
    Acked-by: Ian Rogers <irogers@google.com>
    Cc: Hao Luo <haoluo@google.com>
    Cc: Song Liu <song@kernel.org>
    Cc: bpf@vger.kernel.org
    Link: https://lore.kernel.org/r/20231020204741.1869520-2-namhyung@kernel.org

Signed-off-by: Michael Petlan <mpetlan@redhat.com>