perf tests: test_arm_coresight: Simplify source iteration
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit 0b783d2e82d827af73c779e8e8f95d07e992b451
Author: James Clark <james.clark@arm.com>
Date: Mon Oct 23 14:15:49 2023 +0100

description
===========
There are two reasons to do this, firstly there is a shellcheck warning
in cs_etm_dev_name(), which can be completely deleted. And secondly the
current iteration method doesn't support systems with both ETE and ETM
because it picks one or the other. There isn't a known system with this
configuration, but it could happen in the future.

Iterating over all the sources for each CPU can be done by going through
/sys/bus/event_source/devices/cs_etm/cpu* and following the symlink back
to the Coresight device in /sys/bus/coresight/devices. This will work
whether the device is ETE, ETM or any future name, and is much simpler
and doesn't require any hard coded version numbers

    Suggested-by: Suzuki K Poulose <suzuki.poulose@arm.com>
    Signed-off-by: James Clark <james.clark@arm.com>
    Acked-by: Ian Rogers <irogers@google.com>
    Tested-by: Leo Yan <leo.yan@linaro.org>
    Cc: tianruidong@linux.alibaba.com
    Cc: Kajol Jain <kjain@linux.ibm.com>
    Cc: Anushree Mathur <anushree.mathur@linux.vnet.ibm.com>
    Cc: Tiezhu Yang <yangtiezhu@loongson.cn>
    Cc: atrajeev@linux.vnet.ibm.com
    Cc: coresight@lists.linaro.org
    Link: https://lore.kernel.org/r/20231023131550.487760-1-james.clark@arm.com
    Signed-off-by: Namhyung Kim <namhyung@kernel.org>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>