thermal: trip: Drop lockdep assertion from thermal_zone_trip_id()
JIRA: https://issues.redhat.com/browse/RHEL-26871

commit 108ffd12be24ba1d74b3314df8db32a0a6d55ba5
Author: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Date:   Wed Oct 11 17:45:42 2023 +0200

The lockdep assertion in thermal_zone_trip_id() triggers when the
trip point sysfs attribute of a thermal instance is read, because
there is no thermal zone locking in that code path.

This is not verly useful, though, because there is no mechanism by which
the location of the trips[] table in a thermal zone or its size can
change after binding cooling devices to the trips in that thermal
zone and before those cooling devices are unbound from them.  Thus
it is not in fact necessary to hold the thermal zone lock when
thermal_zone_trip_id() is called from trip_point_show() and so the
lockdep asserion in the former is invalid.

Accordingly, drop that lockdep assertion.

Fixes: 2c7b4bfadef0 ("thermal: core: Store trip pointer in struct thermal_instance")
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>