perf beauty socket/prctl_option: Cope with extended regexp complaint by grep
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit 1715b6359c1ae37f24d6774f0bcd73b6bf839eaa
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Oct 6 16:14:54 2023 -0300

description
===========
Noticed on fedora 38, the extended regexp that so far was ok for both
grep and sed now gets complaints by grep, that says '/' doesn't need to
be escaped with '\'.

So stop using '/' in sed, use '%' instead and remove the \ before / in
the common extended regexp.

Link: https://x.com/SMT_Solvers/status/1710380010098344192?s=20
    Cc: Adrian Hunter <adrian.hunter@intel.com>
    Cc: Ian Rogers <irogers@google.com>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Cc: Namhyung Kim <namhyung@kernel.org>
    Link: https://lore.kernel.org/lkml/ZUEddFPTJHVLhH%2F6@kernel.org
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>