perf test sigtrap: Generalize the BTF routine to reuse it in this test
JIRA: https://issues.redhat.com/browse/RHEL-25824

upstream
========
commit a472ee42e6f60c8714e2306385687922afcba8c4
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Nov 29 12:47:17 2023 -0300

description
===========
Move the part that loads the BTF info to a "btf__available()" that will
lazy load the BTF info so that if we need it for some other test, which
we will in the following cset, we can reuse it.

At some point this will move from this specific 'perf test' entry to be
used in other parts of perf, do it when needed.

    Cc: Adrian Hunter <adrian.hunter@intel.com>
    Cc: Clark Williams <williams@redhat.com>
    Cc: Ian Rogers <irogers@google.com>
    Cc: Ingo Molnar <mingo@kernel.org>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Cc: Kate Carcia <kcarcia@redhat.com>
    Cc: Namhyung Kim <namhyung@kernel.org>
    Cc: Thomas Gleixner <tglx@linutronix.de>
    Link: https://lore.kernel.org/r/20231129154718.326330-2-acme@kernel.org
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>